feat : getInstance 삭제(annotation 활용으로 대체)
